Adds check for database existing; adds setup for /plots folder; updates messages.

+++ cluster_shelving.sh
@@ -5,17 +5,21 @@
 database="cluster_analysis"
 
 # Set up an empty database with the name above.
-sudo -u $USER dropdb --if-exists $database
-sudo -u $USER createdb $database
-
-# Status update.
-echo "$database dropped and (re)created."
+# Drop the existing database if it exists.
+if psql $database -c '\q' 2>&1; then
+	sudo -u $USER dropdb --if-exists $database
+	sudo -u $USER createdb $database
+	echo "$database dropped and recreated."
+else
+	sudo -u $USER createdb $database
+	echo "$database created."
+fi
 
 # Create the table in the database to hold the LibraryThing dump file's data.
 psql $database -f create_table.sql > /dev/null 2>&1
 
 # Status update.
-echo "Table created in $database."
+echo "public.library created in $database."
 
 # Load the LibraryThing dump file (converted to CSV) into the database and
 # record the number of rows imported.
@@ -25,7 +29,7 @@ n_records_loaded=$(psql $database -f load_table.sql)
 n_records_loaded=${n_records_loaded//[a-zA-Z ]/}
 
 # Status update.
-echo "${n_records_loaded} records from LibraryThing dump file loaded into $database."
+echo "${n_records_loaded} records from LibraryThing dump file loaded into public.library."
 
 # Clean up the height into a standard format and Imperial units and record the
 # number of non-NULL values that are produced.
@@ -74,3 +78,13 @@ n_no_dimensions=${n_no_dimensions// /}
 
 # Status update.
 echo "${n_dimensions} records with height/length/thickness, ${n_no_dimensions} records without."
+
+# Get ready to output first-pass summary plots.
+# Clear out the /plots directory (if it exists).
+[ -d plots ] && { rm -rf plots; echo "Existing /plots directory removed."; }
+
+# Create the /plots directory.
+mkdir plots
+
+# Status update.
+echo "Created /plots directory."
